Add files via upload
The secret messages have been very crucial to Secret Agencies. But the most amazing fact is there are lots more ways to make it. Here is one of the important examples described with the code. Also here is the link for the app known as "Cryptech" to use. Just do certain processes to use it. You have to manually give permission to storage from the Settings>Applications>Cryptech>Permissions>Storage>Allow. Enjoy it!.